Drivers: misc: remove __dev* attributes.
CONFIG_HOTPLUG is going away as an option.  As a result, the __dev*
markings need to be removed.

This change removes the use of __devinit, __devexit_p, __devinitdata,
__devinitconst, and __devexit from these drivers.

Based on patches originally written by Bill Pemberton, but redone by me
in order to handle some of the coding style issues better, by hand.

Cc: Bill Pemberton <wfp5p@virginia.edu>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

Conflicts:
	drivers/bus/omap-ocp2scp.c
	drivers/bus/omap_l3_noc.c
	drivers/gpio/gpio-da9055.c
	drivers/gpio/gpio-ts5500.c
	drivers/gpio/gpio-viperboard.c
	drivers/hid/i2c-hid/i2c-hid.c
	drivers/hwmon/vexpress.c
	drivers/input/keyboard/tca8418_keypad.c
	drivers/nfc/pn544/i2c.c
	drivers/sh/pfc/gpio.c
	drivers/sh/pfc/pinctrl.c