clk: qcom: Increase branch clk status maximum delay
In most situations, the status bit of a branch clock is polled until it
reaches the desired on/off status. However, in the case where the
parent clock may be turned off without warning, this status bit no
longer indicates the correct status. Use the same maximum delay value
for these two situations.

Signed-off-by: Patrick Daly <pdaly@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>