usb: xhci: Release spinlock during command cancellation
When a command fails due to timeout, during the command
cancellation handled in xhci_cancel_cmd() the call to
xhci_abort_cmd_ring() is called while the xhci spinlock
is held and with interrupts disabled. However, if the
abortion itself fails because of a dead host controller,
it may take up to 5 seconds before it returns, during
which time a watchdog will time out. Fix this by releasing
the spin lock prior to attempting to abort the command ring.

Change-Id: I3be0edbbd8a6f1c39ef0c563b10de050527e97e9
Signed-off-by: Jack Pham <jackp@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>